Refactorizando estilos y arquitectura de archvios de la app
