+++ src/Components/FavoritesSection/Favorites.css
