--- src/Components/Icons.css
@@ -1,4 +1,0 @@
-.fav-icon{
-    width: 25px;
-    height: 25px;
-}
