+++ src/Components/Icons/Icons.jsx
@@ -1,4 +1,3 @@
-import './Icons.css'
 
 export function AddToFavorite () {
   return (
